Renamed LayerPicker files to more convenient names
